testutil: move file presence checker to separate file
