+++ testutil/checkers.go
@@ -23,45 +23,12 @@ import (
 	"bytes"
 	"fmt"
 	"io/ioutil"
-	"os"
 	"reflect"
 	"regexp"
 	"strings"
 
 	"gopkg.in/check.v1"
 )
-
-type filePresenceChecker struct {
-	*check.CheckerInfo
-	present bool
-}
-
-// FilePresent verifies that the given file exists.
-var FilePresent check.Checker = &filePresenceChecker{
-	CheckerInfo: &check.CheckerInfo{Name: "FilePresent", Params: []string{"filename"}},
-	present:     true,
-}
-
-// FileAbsent verifies that the given file does not exist.
-var FileAbsent check.Checker = &filePresenceChecker{
-	CheckerInfo: &check.CheckerInfo{Name: "FileAbsent", Params: []string{"filename"}},
-	present:     false,
-}
-
-func (c *filePresenceChecker) Check(params []interface{}, names []string) (result bool, error string) {
-	filename, ok := params[0].(string)
-	if !ok {
-		return false, "filename must be a string"
-	}
-	_, err := os.Stat(filename)
-	if os.IsNotExist(err) && c.present {
-		return false, fmt.Sprintf("file %q is absent but should exist", filename)
-	}
-	if err == nil && !c.present {
-		return false, fmt.Sprintf("file %q is present but should not exist", filename)
-	}
-	return true, ""
-}
 
 type fileContentChecker struct {
 	*check.CheckerInfo
